health actuator for edc dummy server
