+++ dataland-backend/src/main/kotlin/org/dataland/datalandbackend/health/EdcDummyServerHealthIndicator.kt
@@ -1,0 +1,6 @@
+package org.dataland.datalandbackend.health
+
+import org.springframework.stereotype.Component
+
+@Component
+class EdcDummyServerHealthIndicator : DummyServerHealthIndicator("http://edc-dummyserver:8080/actuator/health")
