+++ dataland-backend/src/test/kotlin/org/dataland/datalandbackend/health/EdcDummyServerHealthIndicatorTest.kt
@@ -1,0 +1,31 @@
+package org.dataland.datalandbackend.health
+
+import org.junit.jupiter.api.Assertions.assertEquals
+import org.junit.jupiter.api.Assertions.assertFalse
+import org.junit.jupiter.api.Test
+import org.springframework.boot.test.context.SpringBootTest
+
+@SpringBootTest
+class EdcDummyServerHealthIndicatorTest {
+    val edcDummyServerHealthIndicator = EdcDummyServerHealthIndicator()
+
+    @Test
+    fun `test if health of the Skyminder Dummy Server has the status DOWN because it is not responding`() {
+        val status: String = edcDummyServerHealthIndicator.health().status.code
+        assertEquals(
+            status, "DOWN",
+            "The Skyminder Dummy Server Health Indicator is not returning the status \"DOWN\"." +
+                    " It should return \"DOWN\", since this is a unit test and no services should be reachable."
+        )
+    }
+
+    @Test
+    fun `test if the Skyminder Dummy Server container is not responding and therefore the output is false`() {
+        val isRunningSkyminderServer: Boolean = edcDummyServerHealthIndicator.isRunningSkyminderServer()
+        assertFalse(
+            isRunningSkyminderServer,
+            "The Skyminder Dummy Server Health Indicator is not returning the status \"DOWN\"." +
+                    " It should return \"DOWN\", since this is a unit test and no services should be reachable."
+        )
+    }
+}
